Remove subscription when there's an error while sending the notification

+++ wp-web-push/wp-web-push-main.php
@@ -226,20 +226,15 @@ class WebPush_Main {
         continue;
       }
 
-      // TODO: Clean approximately ten random subscriptions, to avoid performance problems
-      // with sending too many synchronous requests.
-      //$sync = mt_rand(1, $subscription_num) <= 10;
-
-      /*if (!sendNotification($subscription->endpoint, $isGCM, $gcmKey, $sync)) {
-        // If there's an error while sending the push notification,
-        // the subscription is no longer valid, hence we remove it.
-        WebPush_DB::remove_subscription($subscription->endpoint);
-      } else {
-        $notification_count++;
-      }*/
-
-      $webPush->addRecipient($subscription->endpoint, $isGCM, $gcmKey);
-      $notification_count++;
+      $webPush->addRecipient($subscription->endpoint, $isGCM, $gcmKey, function($success) use ($subscription, &$notification_count) {
+        if (!$success) {
+          // If there's an error while sending the push notification,
+          // the subscription is no longer valid, hence we remove it.
+          WebPush_DB::remove_subscription($subscription->endpoint);
+        } else {
+          $notification_count++;
+        }
+      });
     }
 
     $webPush->sendNotifications();
